タグ追加＆花火発見機能分離
