+++ firework_game/ProjectSettings/TagManager.asset
@@ -5,6 +5,8 @@ TagManager:
   serializedVersion: 3
   tags:
   - Ground
+  - NPC
+  - Police
   layers:
   - Default
   - TransparentFX
